requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0